+++ example/lib/src/core/router/router.dart
@@ -17,18 +17,9 @@ class AppRouter extends RootStackRouter {
       canOnlySwipeFromEdge: true,
     ),
     SwiftSheetAutoRoute(page: SheetRoute.page),
-    SwiftSheetAutoRoute(
-      page: SheetNoBgRoute.page,
-      animateBackground: false,
-    ),
-    SwiftSheetAutoRoute(
-      page: SheetNoSwipeRoute.page,
-      enableDrag: false,
-    ),
-    SwiftSheetAutoRoute(
-      page: SheetCustomRadiusRoute.page,
-      sheetRadius: 16,
-    ),
+    SwiftSheetAutoRoute(page: SheetNoBgRoute.page, animateBackground: false),
+    SwiftSheetAutoRoute(page: SheetNoSwipeRoute.page, enableDrag: false),
+    SwiftSheetAutoRoute(page: SheetCustomRadiusRoute.page, sheetRadius: 16),
     SwiftPageAutoRoute(page: HeroRouteOneRoute.page),
     SwiftPageAutoRoute(page: HeroRouteTwoRoute.page),
     SwiftSheetAutoRoute(page: HeroSheetOneRoute.page),