+++ example_go_router/lib/src/screens/splash_screen.dart
@@ -1,0 +1,18 @@
+import 'package:flutter/material.dart';
+import 'package:go_router/go_router.dart';
+
+class SplashScreen extends StatelessWidget {
+  const SplashScreen({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      body: Center(
+        child: FilledButton(
+          onPressed: () => context.push('/home'),
+          child: const Text('Enter'),
+        ),
+      ),
+    );
+  }
+}